Modified audioDownloaderclass to fit the needs of "After voiceover"
Created videoBuilder class which instantiates getTitle class and allows me to build a dynamic video with a new backdrop every time when it asks for the title.

I now have a way of going from voiceover to video.

Now I just need to figure out how to auto post and I'm going to build a class that will do just that.